# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Address API Collect&Verything",
        "version": "0.1"
    },
    "paths": {
        "/api/addresses/{id}": {
            "get": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Get address information",
                "description": "Returns address data",
                "operationId": "retrieve",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Address id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Account not found."
                    },
                    "409": {
                        "description": "Conflict"
                    },
                    "500": {
                        "description": "Servor Error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Delete a address",
                "description": "Soft delete a address",
                "operationId": "delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Account id",
                        "required": true,
                        "schema": {
                            "type": "String"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Account deleted"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Patch a address",
                "description": "Update an address",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Address title"
                    },
                    {
                        "name": "address_line_1",
                        "in": "query",
                        "description": "Address line 1"
                    },
                    {
                        "name": "address_line_2",
                        "in": "query",
                        "description": "Address line 2"
                    },
                    {
                        "name": "zip_code",
                        "in": "query",
                        "description": "Zip code"
                    },
                    {
                        "name": "city",
                        "in": "query",
                        "description": "City"
                    },
                    {
                        "name": "country",
                        "in": "query",
                        "description": "Address country"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Account updated"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/addresses": {
            "get": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Get all addresses information",
                "description": "Returns address data",
                "operationId": "list",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "409": {
                        "description": "Conflict"
                    },
                    "500": {
                        "description": "Servor Error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Post a new address",
                "description": "Create a new address",
                "operationId": "create",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Address title",
                        "required": true
                    },
                    {
                        "name": "address_line_1",
                        "in": "query",
                        "description": "Address line 1",
                        "required": true
                    },
                    {
                        "name": "address_line_2",
                        "in": "query",
                        "description": "Address line 2"
                    },
                    {
                        "name": "zip_code",
                        "in": "query",
                        "description": "Zip code",
                        "required": true
                    },
                    {
                        "name": "city",
                        "in": "query",
                        "description": "City",
                        "required": true
                    },
                    {
                        "name": "country",
                        "in": "query",
                        "description": "Address country",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Account created"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}